ci: Switch to ARM64 runner
As a result of internal test, arm64 runner is faster than x86_64 runner.
Thus, we change to use arm64 runner.

Note:
ARM64 runner is provided under **public preview** as of now.
The biggest issue is /mnt is not provided so that arm64 runner has about
only 50GB build space. To build with --rm_work option, we can build image now.

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>